Update message for use by MetaMask
The PR message for this bot was not appropriate for use by MetaMask
because it implied that the project was open source, and because it
assumed the signature would be applicable for all ConsenSys projects.

Both issues have now been addressed. There is no longer any implication
that the project is open source, and it's now clear that the signature
applies only to this repository.